Updating 'depends_on macos' lines for new formatting
When running `brew cleanup`, I saw the following message:
Warning: Calling depends_on macos: "10.11" is deprecated and will be disabled on 2019-10-15! Use depends_on macos: :el_capitan instead.
Please report this to the puppetlabs/puppet tap, or even better, submit a PR to fix it:
  /usr/local/Homebrew/Library/Taps/puppetlabs/homebrew-puppet/Casks/pdk.rb:21

So here's the commit to fix it.