Add a recommended dependency on coreutils
This installs `realpath`, which Wash uses to provide its shell prompt.
Coreutils will be installed by default, but use `recommended` to provide
the option to install without it because `coreutils` brings in several
applications that conflict with other packages. A future version of Wash
will include a fix for https://github.com/puppetlabs/wash/issues/492
that degrades gracefully if `realpath` is not present.

Signed-off-by: Michael Smith <michael.smith@puppet.com>