(maint) Update rake task to include macOS 11 package for puppet-bolt
This updates the rake task to include the macOS 11 package for
puppet-bolt. Previously, the puppet-tools collection was used to
determine the OS versions to include in the puppet-bolt cask. However,
the puppet-tools collection includes pdk, which does not ship for macOS
11 currently, necessitating additional logic to not include macOS 11 in
the list of OS versions for any puppet-tools package other than
puppet-bolt.