(maint) Guard cask:brew task arm64 lookup for non puppet-agent
b92dab474 updated the Rakefile to lookup arm64 shas for puppet-agent.
Bolt doesn't have arm64 packages (BOLT-1604 is pending). Right now only
puppet-agent has arm64 packages so added a check for pkg to skip
attempting to fetch arm64 if not puppet-agent.

The templates for pdk and pe-client-tools probably need to be updated as
well so that they can handle the addition of 'arch' in stanza and url,
but I'll open a separate pr for that so as not to hold up the
puppet-bolt release further.